+++ zeroeq/publisher.cpp
@@ -5,7 +5,7 @@
  */
 
 #include "publisher.h"
-#include "event.h"
+#include "fbevent.h"
 #include "log.h"
 #include "detail/broker.h"
 #include "detail/byteswap.h"